complete modularization
