+++ src/util.h
@@ -1,5 +1,7 @@
 #include <fstream>
-#include "grid.h"
+#ifndef GRID_H
+#define GRID_H
+#endif 
 
 
 void file_to_grids(std::string filename, std::vector<Grid> &grids){
